requirements.txt
langchain==0.3.23
langchain_community==0.3.21
nltk==3.9.1
numpy==2.2.4
python-dotenv==1.1.0
scikit_learn==1.6.1
streamlit==1.44.1
huggingface_hub==0.30.2
sentence-transformers==2.2.2
faiss-cpu==1.10.0